refactor(edge-worker): apply SOLID to memory-limit hooks
Address review feedback. Restructure the two hooks for clearer
single-responsibility and dependency-inversion, mirroring the existing
PrMarkerProvider / IntentToAddGitClient seams:

- Extract hooks/cyrus-tool-exec.ts as the single source of truth for the
  wrapper CLI contract (path, env var, OOM marker, wrap/unwrap, parse).
  Removes the wrapped-command-format coupling that was duplicated across
  the producing (PreToolUse) and consuming (OOM) hooks (SRP/DRY).
- Split the OOM hook into hooks/OomReportHook.ts and introduce an
  OomEventReporter interface with a default HttpOomEventReporter. The
  hook now only detects+parses OOMs and delegates delivery to the
  injected reporter; transport details (auth/URL/timeout/fail-open) live
  in the reporter (DIP/SRP/OCP).
- MemoryLimitHook.ts is now PreToolUse-only and delegates quoting/format
  to wrapCommand.
- Split tests into cyrus-tool-exec / MemoryLimitHook / OomReportHook
  suites (40 tests); reporter gating now tested at the seam.
