+++ packages/edge-worker/test/MemoryLimitHook.test.ts
@@ -1,19 +1,11 @@
 import type {
 	HookCallbackMatcher,
-	PostToolUseHookInput,
 	PreToolUseHookInput,
 } from "cyrus-claude-runner";
 import type { ILogger } from "cyrus-core";
-import { describe, expect, it, vi } from "vitest";
-import {
-	buildCommandExcerpt,
-	buildMemoryLimitHook,
-	buildOomReportHook,
-	extractResultText,
-	OOM_MARKER,
-	parseOomMarker,
-	singleQuote,
-} from "../src/hooks/MemoryLimitHook.js";
+import { describe, expect, it } from "vitest";
+import { wrapCommand } from "../src/hooks/cyrus-tool-exec.js";
+import { buildMemoryLimitHook } from "../src/hooks/MemoryLimitHook.js";
 
 const silentLogger: ILogger = {
 	debug: () => {},
@@ -65,20 +57,6 @@ function preMatcher(
 	return matcher;
 }
 
-describe("singleQuote", () => {
-	it("wraps a simple string", () => {
-		expect(singleQuote("echo hi")).toBe("'echo hi'");
-	});
-
-	it("escapes embedded single quotes", () => {
-		expect(singleQuote("it's")).toBe("'it'\\''s'");
-	});
-
-	it("leaves double quotes, backticks and $() untouched inside quotes", () => {
-		expect(singleQuote('echo "$(date)" `id`')).toBe("'echo \"$(date)\" `id`'");
-	});
-});
-
 describe("buildMemoryLimitHook — gating (no-op cases)", () => {
 	it("registers a Bash matcher under PreToolUse", () => {
 		const matcher = preMatcher({});
@@ -126,7 +104,7 @@ describe("buildMemoryLimitHook — command rewrite", () => {
 			hookEventName: "PreToolUse",
 			permissionDecision: "allow",
 			updatedInput: {
-				command: "CYRUS_TOOL_MEMORY_MAX_MB=1300 cyrus-tool-exec 'pnpm test'",
+				command: wrapCommand("pnpm test", "1300"),
 			},
 		});
 	});
@@ -138,7 +116,7 @@ describe("buildMemoryLimitHook — command rewrite", () => {
 			makePreInput({ command: "ls", description: "list", timeout: 5000 }),
 		);
 		expect(result.hookSpecificOutput.updatedInput).toEqual({
-			command: "CYRUS_TOOL_MEMORY_MAX_MB=1300 cyrus-tool-exec 'ls'",
+			command: wrapCommand("ls", "1300"),
 			description: "list",
 			timeout: 5000,
 		});
@@ -146,12 +124,10 @@ describe("buildMemoryLimitHook — command rewrite", () => {
 
 	it("safely single-quotes embedded single quotes", async () => {
 		const matcher = preMatcher(deps);
-		const result = await runPre(
-			matcher,
-			makePreInput({ command: "echo 'hello world'" }),
-		);
+		const cmd = "echo 'hello world'";
+		const result = await runPre(matcher, makePreInput({ command: cmd }));
 		expect(result.hookSpecificOutput.updatedInput.command).toBe(
-			"CYRUS_TOOL_MEMORY_MAX_MB=1300 cyrus-tool-exec 'echo '\\''hello world'\\'''",
+			wrapCommand(cmd, "1300"),
 		);
 	});
 
@@ -160,7 +136,7 @@ describe("buildMemoryLimitHook — command rewrite", () => {
 		const cmd = 'echo "$(whoami)" `hostname`';
 		const result = await runPre(matcher, makePreInput({ command: cmd }));
 		expect(result.hookSpecificOutput.updatedInput.command).toBe(
-			`CYRUS_TOOL_MEMORY_MAX_MB=1300 cyrus-tool-exec ${singleQuote(cmd)}`,
+			wrapCommand(cmd, "1300"),
 		);
 	});
 
@@ -169,7 +145,7 @@ describe("buildMemoryLimitHook — command rewrite", () => {
 		const cmd = "cat <<'EOF'\nline1\nline2\nEOF";
 		const result = await runPre(matcher, makePreInput({ command: cmd }));
 		expect(result.hookSpecificOutput.updatedInput.command).toBe(
-			`CYRUS_TOOL_MEMORY_MAX_MB=1300 cyrus-tool-exec ${singleQuote(cmd)}`,
+			wrapCommand(cmd, "1300"),
 		);
 	});
 });
@@ -199,149 +175,5 @@ describe("buildMemoryLimitHook — fail open on unexpected input", () => {
 		const matcher = preMatcher(deps);
 		const result = await runPre(matcher, makePreInput({ command: "" }));
 		expect(result).toEqual({ continue: true });
-	});
-});
-
-describe("parseOomMarker", () => {
-	it("parses cap and peak from the marker line", () => {
-		const line = `${OOM_MARKER} exceeded 1300M memory budget (peak 1500000000 bytes).`;
-		expect(parseOomMarker(line)).toEqual({
-			budgetMb: 1300,
-			peakBytes: 1500000000,
-		});
-	});
-
-	it("returns an empty object when nothing matches", () => {
-		expect(parseOomMarker("no marker here")).toEqual({});
-	});
-});
-
-describe("extractResultText", () => {
-	it("returns strings unchanged", () => {
-		expect(extractResultText("hello")).toBe("hello");
-	});
-
-	it("joins string fields of an object (e.g. stdout/stderr)", () => {
-		const text = extractResultText({ stdout: "out", stderr: "err" });
-		expect(text).toContain("out");
-		expect(text).toContain("err");
-	});
-});
-
-describe("buildCommandExcerpt", () => {
-	it("returns a short prefix of a plain command", () => {
-		expect(buildCommandExcerpt("pnpm test")).toBe("pnpm test");
-	});
-
-	it("unwraps the wrapper prefix back to the original command", () => {
-		const wrapped =
-			"CYRUS_TOOL_MEMORY_MAX_MB=1300 cyrus-tool-exec 'echo '\\''hi'\\'''";
-		expect(buildCommandExcerpt(wrapped)).toBe("echo 'hi'");
-	});
-
-	it("truncates to the max length", () => {
-		const long = "a".repeat(500);
-		expect(buildCommandExcerpt(long, 200)).toHaveLength(200);
-	});
-});
-
-function makePostInput(
-	toolResponse: unknown,
-	command = "pnpm test",
-): PostToolUseHookInput {
-	return {
-		hook_event_name: "PostToolUse",
-		session_id: "s",
-		transcript_path: "t",
-		cwd: "/work",
-		tool_name: "Bash",
-		tool_input: { command },
-		tool_response: toolResponse,
-		tool_use_id: "u",
-	} as PostToolUseHookInput;
-}
-
-async function runPost(
-	matcher: HookCallbackMatcher,
-	input: PostToolUseHookInput,
-): Promise<any> {
-	const fn = matcher.hooks[0];
-	return fn(input as any, "u", { signal: new AbortController().signal });
-}
-
-function postMatcher(
-	deps: Parameters<typeof buildOomReportHook>[1],
-): HookCallbackMatcher {
-	const hook = buildOomReportHook(silentLogger, deps);
-	const matcher = hook.PostToolUse?.[0];
-	if (!matcher) {
-		throw new Error("expected a PostToolUse matcher");
-	}
-	return matcher;
-}
-
-describe("buildOomReportHook", () => {
-	const markerText = `${OOM_MARKER} exceeded 1300M memory budget (peak 1500000000 bytes).`;
-
-	it("registers a Bash matcher under PostToolUse", () => {
-		const matcher = postMatcher({ getEnv: () => undefined });
-		expect(matcher.matcher).toBe("Bash");
-	});
-
-	it("does not POST when the marker is absent", async () => {
-		const fetchImpl = vi.fn();
-		const matcher = postMatcher({
-			getEnv: cloudEnv({ CYRUS_API_KEY: "key" }),
-			fetchImpl: fetchImpl as unknown as typeof fetch,
-		});
-		await runPost(matcher, makePostInput("all good"));
-		expect(fetchImpl).not.toHaveBeenCalled();
-	});
-
-	it("does not POST when CYRUS_API_KEY is missing", async () => {
-		const fetchImpl = vi.fn();
-		const matcher = postMatcher({
-			getEnv: () => undefined,
-			fetchImpl: fetchImpl as unknown as typeof fetch,
-		});
-		await runPost(matcher, makePostInput(markerText));
-		expect(fetchImpl).not.toHaveBeenCalled();
-	});
-
-	it("POSTs the parsed OOM event to /api/oom-event with the bearer key", async () => {
-		const fetchImpl = vi.fn().mockResolvedValue({ ok: true, status: 200 });
-		const matcher = postMatcher({
-			getEnv: (n) => (n === "CYRUS_API_KEY" ? "secret-key" : undefined),
-			getBaseUrl: () => "https://app.atcyrus.com",
-			fetchImpl: fetchImpl as unknown as typeof fetch,
-		});
-		await runPost(
-			matcher,
-			makePostInput({ stderr: markerText }, "pnpm run heavy"),
-		);
-
-		expect(fetchImpl).toHaveBeenCalledTimes(1);
-		const [url, options] = fetchImpl.mock.calls[0];
-		expect(url).toBe("https://app.atcyrus.com/api/oom-event");
-		expect(options.method).toBe("POST");
-		expect(options.headers.Authorization).toBe("Bearer secret-key");
-		expect(JSON.parse(options.body)).toEqual({
-			budgetMb: 1300,
-			peakBytes: 1500000000,
-			commandExcerpt: "pnpm run heavy",
-		});
-	});
-
-	it("fails open (returns {}) when fetch throws", async () => {
-		const fetchImpl = vi.fn().mockRejectedValue(new Error("network down"));
-		const matcher = postMatcher({
-			getEnv: (n) => (n === "CYRUS_API_KEY" ? "secret-key" : undefined),
-			fetchImpl: fetchImpl as unknown as typeof fetch,
-		});
-		const result = await runPost(
-			matcher,
-			makePostInput({ stderr: markerText }),
-		);
-		expect(result).toEqual({});
 	});
 });
